refactor: axios 기반 회고 API 호출 로직 도입 및 타입 분리 리팩토링
