+++ src/apis/retrospective.ts
@@ -1,32 +1,6 @@
 import axios from 'axios';
 
-export interface Question {
-  questionId: number;
-  category: string;
-  content: string;
-  isSelected: boolean;
-  answerId: number | null;
-  answer: string | null;
-}
-
-export interface PullRequestDetail {
-  id: number;
-  title: string;
-  tag: string;
-  recordStatus: 'PENDING' | 'PROGRESS' | 'DONE';
-  mergedAt: string;
-  summary: string;
-  categories: string[];
-  questions: Question[];
-}
-
-export interface CategoryWithQuestions {
-  category: string;
-  questions: {
-    questionId: number;
-    question: string;
-  }[];
-}
+import type { PullRequestDetail } from '@/types/retrospective';
 
 export async function fetchPullRequestById(pullRequestId: string, accessToken: string): Promise<PullRequestDetail> {
   try {
